+++ lib/src/custom_sentence.dart
@@ -1,21 +1,15 @@
-import 'package:nmea/src/checksum_sentence.dart';
-import 'package:nmea/src/nmea_sentence_type.dart';
+import 'package:nmea/nmea.dart';
 
-/// The data format is completely customizable, as long as it fits in the general
-/// envelope of an NMEA0183 sentence (begins with '$' and ends with windows-
-/// style newline characters).
-class CustomSentence extends ChecksumSentence {
+/// An [NmeaSentence] where the data format is completely customizable, as
+/// long as it fits in the general envelope of an NMEA0183 sentence (begins with
+/// '$' and ends with windows-style newline characters).
+class CustomSentence extends NmeaSentence {
   /// Returns the identifier used to identify this sentence type
   final String identifier;
-  final bool validateChecksums;
 
-  CustomSentence(
-      {required this.identifier,
-      required super.raw,
-      super.prefix,
-      this.validateChecksums = true})
-      : super(type: NmeaSentenceType.unknown);
-
-  @override
-  bool get hasValidChecksum => super.hasValidChecksum || !validateChecksums;
+  /// The [CustomChecksumSentence] constructor forwards all arguments to the
+  /// [NmeaSentence] constructor, except the [type] parameter.
+  /// The [type] is always [NmeaSentenceType.custom].
+  CustomSentence({required this.identifier, required super.raw, super.prefix})
+      : super(type: NmeaSentenceType.custom);
 }